+++ core/site.ts
@@ -453,6 +453,12 @@ export default class Site {
     return this;
   }
 
+  /** Copy the remaining files */
+  copyRemainingFiles(filter: (path: string) => string | boolean): this {
+    this.source.copyRemainingFiles = filter;
+    return this;
+  }
+
   /** Ignore one or several files or directories */
   ignore(...paths: (string | ScopeFilter)[]): this {
     paths.forEach((path) => {